+++ CHANGELOG.md
@@ -48,6 +48,14 @@ This file is used to list changes made in each version of the managed-chef-serve
 
 - policyfile_loader now puts policyfiles in a _default policygroup as defined by an attribute.
 
+# 0.7.0
+
+- data_bag_loader recipe and tests
+
 # BACKLOG
 
 - maintenance tasks
+sudo /opt/chef/embedded/bin/inspec exec inspec-chef-server --attrs=config.yml
+sudo /opt/chef/embedded/bin/inspec exec https://github.com/chef/inspec-chef-server.git --attrs=config.yml
+inspec exec https://github.com/mattray/inspec-chef-server/tree/rhel --attrs=config.yml
+inspec exec https://github.com/mattray/inspec-chef-server/tree/rhel --target=ssh://192.168.33.22:2222 --user=vagrant --key-files=~/.vagrant.d/insecure_private_key --attrs=test/config.yml --sudo